requirements.txt
pyzmq==25.1.2
psutil==5.9.8
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
streamlit==1.32.2
plotly==5.19.0
rich==13.7.1